requirements.txt
Django==1.6
django-uuidfield==0.4.0
South==0.8.4
-e git+https://github.com/dotcloud/docker-py.git#egg=docker-py
djangorestframework==2.3.9